Rename registry package to backendregistry
Addresses stacklok/toolhive#5542 review comments:
- MEDIUM pkg/vmcp/registry/registry.go (3429022196): package name
  collided with the top-level pkg/registry (package registry, 24
  importers) and pkg/vmcp/registry.go (package vmcp). Rename to
  pkg/vmcp/backendregistry (package backendregistry) while there are no
  importers, so dual-importers no longer need an alias.
